requirements.txt
glfw==1.7.1
llvmlite==0.27.0
numba==0.42.0
numpy==1.16.1
numpy-quaternion==2019.2.12.9.32.25
PyOpenGL==3.1.0
